handshake: cast LAP_TIME_INVALID to int32_t in the l2 sector ternary
best_sectors_ms is int32_t but LAP_TIME_INVALID is 0x7FFFFFFFu, so the ?:
result was unsigned (-Wsign-compare).  0x7FFFFFFF = INT32_MAX fits int32_t
exactly; cast it so both branches share signedness and the build is clean.

+++ accd/handshake.c
@@ -1188,7 +1188,7 @@ write_car_leaderboard_record(struct ByteBuf *bb,
 			for (k = 0; k < 3; k++)
 				l2_buf[k] = race->best_sectors_ms[k] > 0
 				    ? race->best_sectors_ms[k]
-				    : LAP_TIME_INVALID;
+				    : (int32_t)LAP_TIME_INVALID;
 		}
 
 		/*
